+++ utils/helpers/helpers.go
@@ -1,15 +1,14 @@
 package helpers
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"math"
 	"net/http"
-	"net/url"
 	"os"
 	"regexp"
+	"stockbackend/clients/http_client"
 	"stockbackend/types"
 	"strconv"
 	"strings"
@@ -419,54 +418,6 @@ func FetchPeerData(dataWarehouseID string) ([]map[string]string, error) {
 	return peersData, nil
 }
 
-func SearchCompany(queryString string) ([]types.Company, error) {
-	// Replace "corporation" with "Corpn" and "limited" with "Ltd"
-	queryString = strings.ReplaceAll(queryString, " Corporation ", " Corpn ")
-	queryString = strings.ReplaceAll(queryString, " corporation ", " Corpn ")
-	queryString = strings.ReplaceAll(queryString, " Limited", " Ltd ")
-	queryString = strings.ReplaceAll(queryString, " limited", " Ltd ")
-	queryString = strings.ReplaceAll(queryString, " and ", " & ")
-	queryString = strings.ReplaceAll(queryString, " And ", " & ")
-	// Base URL for the Screener API
-	baseURL := os.Getenv("COMPANY_URL") + "/api/company/search/"
-
-	// Create the URL with query parameters
-	params := url.Values{}
-	params.Add("q", queryString)
-	params.Add("v", "3")
-	params.Add("fts", "1")
-
-	// Create the request
-	req, err := http.NewRequest("GET", baseURL+"?"+params.Encode(), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	// Create the client and send the request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	// Read the response
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var searchResponse []types.Company
-	err = json.Unmarshal(body, &searchResponse)
-	if err != nil {
-		zap.L().Error("Failed to unmarshal search response", zap.Error(err))
-		return nil, err
-	}
-
-	// Return the list of results
-	return searchResponse, nil
-}
-
 func ParseTableData(section *goquery.Selection, tableSelector string) map[string]interface{} {
 	table := section.Find(tableSelector)
 	if table.Length() == 0 {
@@ -549,18 +500,13 @@ func ParseTable(tableDiv *goquery.Selection) []map[string]interface{} {
 }
 
 func FetchCompanyData(url string) (map[string]interface{}, error) {
-	resp, err := http.Get(url)
+	body, err := http_client.GetCompanyPage(url)
 	if err != nil {
-		return nil, fmt.Errorf("failed to fetch the URL: %v", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return nil, fmt.Errorf("failed to retrieve the content, status code: %d", resp.StatusCode)
+		return nil, fmt.Errorf("failed to fetch the company page: %v", err)
 	}
 
 	// Parse the HTML content of the company page
-	doc, err := goquery.NewDocumentFromReader(resp.Body)
+	doc, err := goquery.NewDocumentFromReader(body)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse the HTML content: %v", err)
 	}
